pyproject.toml
[project]
name = "mcp-harness"
version = "0.1.0"
description = "AI-Native Harness Engineering for MCP Server quality assurance"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "httpx>=0.28.0",
    "openai>=1.0.0",
    "anthropic>=0.40.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
mcp-lint = "harness.lint.mcp_schema_lint:main"
mcp-eval = "scripts.run_eval:main"
mcp-optimize = "scripts.run_optimize:main"

[tool.hatch.build.targets.wheel]
packages = ["harness", "scripts"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
